rebased voice commands
